HeroManager 싱글턴에 레벨, 경험치 로직 추가

+++ Unity/Assets/Hero/HeroManagerData.cs
@@ -6,7 +6,7 @@ namespace Hero
     {
         public float maxHealth;
         public float maxStamina;
-        public float initialMoney;
+        public int initialMoney;
         public float hitEffectDuration;
         public float staminaRecoverAmount;
     }
